refactor: DTO 롬복 적용 (#55)
