+++ backend/src/main/java/com/spring/blog/authentication/application/dto/TokenResponseDto.java
@@ -1,23 +1,17 @@
 package com.spring.blog.authentication.application.dto;
 
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class TokenResponseDto {
 
     private String token;
     private String userName;
-
-    private TokenResponseDto() {
-    }
-
-    public TokenResponseDto(String token, String userName) {
-        this.token = token;
-        this.userName = userName;
-    }
-
-    public String getToken() {
-        return token;
-    }
-
-    public String getUserName() {
-        return userName;
-    }
 }
