+++ backend/src/main/java/com/spring/blog/authentication/infrastructure/dto/request/AccessTokenRequestDto.java
@@ -1,7 +1,16 @@
 package com.spring.blog.authentication.infrastructure.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class AccessTokenRequestDto {
 
     private String code;
@@ -11,25 +20,4 @@ public class AccessTokenRequestDto {
 
     @JsonProperty("client_secret")
     private String clientSecret;
-
-    private AccessTokenRequestDto() {
-    }
-
-    public AccessTokenRequestDto(String code, String clientId, String clientSecret) {
-        this.code = code;
-        this.clientId = clientId;
-        this.clientSecret = clientSecret;
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-    public String getClientId() {
-        return clientId;
-    }
-
-    public String getClientSecret() {
-        return clientSecret;
-    }
 }
