+++ backend/src/main/java/com/spring/blog/authentication/presentation/dto/OAuthLoginUrlResponse.java
@@ -1,17 +1,16 @@
 package com.spring.blog.authentication.presentation.dto;
 
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class OAuthLoginUrlResponse {
 
     private String url;
-
-    private OAuthLoginUrlResponse() {
-    }
-
-    public OAuthLoginUrlResponse(String url) {
-        this.url = url;
-    }
-
-    public String getUrl() {
-        return url;
-    }
 }
