+++ backend/src/main/java/com/spring/blog/authentication/presentation/dto/OAuthTokenResponse.java
@@ -1,23 +1,17 @@
 package com.spring.blog.authentication.presentation.dto;
 
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class OAuthTokenResponse {
 
     private String token;
     private String userName;
-
-    private OAuthTokenResponse() {
-    }
-
-    public OAuthTokenResponse(String token, String userName) {
-        this.token = token;
-        this.userName = userName;
-    }
-
-    public String getToken() {
-        return token;
-    }
-
-    public String getUserName() {
-        return userName;
-    }
 }
