+++ backend/src/main/java/com/spring/blog/comment/application/dto/request/CommentDeleteRequestDto.java
@@ -1,23 +1,17 @@
 package com.spring.blog.comment.application.dto.request;
 
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class CommentDeleteRequestDto {
 
     private Long commentId;
     private Long userId;
-
-    private CommentDeleteRequestDto() {
-    }
-
-    public CommentDeleteRequestDto(Long commentId, Long userId) {
-        this.commentId = commentId;
-        this.userId = userId;
-    }
-
-    public Long getCommentId() {
-        return commentId;
-    }
-
-    public Long getUserId() {
-        return userId;
-    }
 }
