+++ backend/src/main/java/com/spring/blog/comment/application/dto/request/CommentEditRequestDto.java
@@ -1,29 +1,18 @@
 package com.spring.blog.comment.application.dto.request;
 
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class CommentEditRequestDto {
 
     private Long commentId;
     private Long userId;
     private String content;
-
-    private CommentEditRequestDto() {
-    }
-
-    public CommentEditRequestDto(Long commentId, Long userId, String content) {
-        this.commentId = commentId;
-        this.userId = userId;
-        this.content = content;
-    }
-
-    public Long getCommentId() {
-        return commentId;
-    }
-
-    public Long getUserId() {
-        return userId;
-    }
-
-    public String getContent() {
-        return content;
-    }
 }
