+++ backend/src/main/java/com/spring/blog/comment/application/dto/request/CommentListRequestDto.java
@@ -1,35 +1,19 @@
 package com.spring.blog.comment.application.dto.request;
 
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class CommentListRequestDto {
 
     private Long postId;
     private Long page;
     private Long size;
     private Long pageBlockCounts;
-
-    private CommentListRequestDto() {
-    }
-
-    public CommentListRequestDto(Long postId, Long page, Long size, Long pageBlockCounts) {
-        this.postId = postId;
-        this.page = page;
-        this.size = size;
-        this.pageBlockCounts = pageBlockCounts;
-    }
-
-    public Long getPostId() {
-        return postId;
-    }
-
-    public Long getPage() {
-        return page;
-    }
-
-    public Long getSize() {
-        return size;
-    }
-
-    public Long getPageBlockCounts() {
-        return pageBlockCounts;
-    }
 }
