+++ backend/src/main/java/com/spring/blog/comment/application/dto/request/CommentReplyRequestDto.java
@@ -1,35 +1,19 @@
 package com.spring.blog.comment.application.dto.request;
 
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class CommentReplyRequestDto {
 
     private Long postId;
     private Long userId;
     private Long commentId;
     private String content;
-
-    private CommentReplyRequestDto() {
-    }
-
-    public CommentReplyRequestDto(Long postId, Long userId, Long commentId, String content) {
-        this.postId = postId;
-        this.userId = userId;
-        this.commentId = commentId;
-        this.content = content;
-    }
-
-    public Long getPostId() {
-        return postId;
-    }
-
-    public Long getUserId() {
-        return userId;
-    }
-
-    public Long getCommentId() {
-        return commentId;
-    }
-
-    public String getContent() {
-        return content;
-    }
 }
