+++ backend/src/main/java/com/spring/blog/comment/application/dto/request/CommentWriteRequestDto.java
@@ -1,29 +1,18 @@
 package com.spring.blog.comment.application.dto.request;
 
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class CommentWriteRequestDto {
 
     private Long postId;
     private Long userId;
     private String content;
-
-    private CommentWriteRequestDto() {
-    }
-
-    public CommentWriteRequestDto(Long postId, Long userId, String content) {
-        this.postId = postId;
-        this.userId = userId;
-        this.content = content;
-    }
-
-    public Long getPostId() {
-        return postId;
-    }
-
-    public Long getUserId() {
-        return userId;
-    }
-
-    public String getContent() {
-        return content;
-    }
 }
