+++ backend/src/main/java/com/spring/blog/comment/presentation/dto/request/CommentWriteRequest.java
@@ -1,17 +1,16 @@
 package com.spring.blog.comment.presentation.dto.request;
 
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class CommentWriteRequest {
 
     private String content;
-
-    private CommentWriteRequest() {
-    }
-
-    public CommentWriteRequest(String content) {
-        this.content = content;
-    }
-
-    public String getContent() {
-        return content;
-    }
 }
