+++ backend/src/main/java/com/spring/blog/post/application/dto/request/PostListRequestDto.java
@@ -1,29 +1,18 @@
 package com.spring.blog.post.application.dto.request;
 
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class PostListRequestDto {
 
     private Long page;
     private Long size;
     private Long pageBlockCounts;
-
-    private PostListRequestDto() {
-    }
-
-    public PostListRequestDto(Long page, Long size, Long pageBlockCounts) {
-        this.page = page;
-        this.size = size;
-        this.pageBlockCounts = pageBlockCounts;
-    }
-
-    public Long getPage() {
-        return page;
-    }
-
-    public Long getSize() {
-        return size;
-    }
-
-    public Long getPageBlockCounts() {
-        return pageBlockCounts;
-    }
 }
