+++ backend/src/main/java/com/spring/blog/post/application/dto/response/PostListResponseDto.java
@@ -4,7 +4,16 @@ import com.spring.blog.common.PageMaker;
 import com.spring.blog.post.domain.Post;
 import java.util.List;
 import java.util.stream.Collectors;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class PostListResponseDto {
 
     private List<SimplePostResponseDto> simplePostResponseDtos;
@@ -12,23 +21,6 @@ public class PostListResponseDto {
     private int endPage;
     private boolean prev;
     private boolean next;
-
-    private PostListResponseDto() {
-    }
-
-    public PostListResponseDto(
-        List<SimplePostResponseDto> simplePostResponseDtos,
-        int startPage,
-        int endPage,
-        boolean prev,
-        boolean next
-    ) {
-        this.simplePostResponseDtos = simplePostResponseDtos;
-        this.startPage = startPage;
-        this.endPage = endPage;
-        this.prev = prev;
-        this.next = next;
-    }
 
     public static PostListResponseDto from(List<Post> posts, PageMaker pageMaker) {
         List<SimplePostResponseDto> postResponseDtos = posts.stream()
@@ -41,25 +33,5 @@ public class PostListResponseDto {
             pageMaker.isPrev(),
             pageMaker.isNext()
         );
-    }
-
-    public List<SimplePostResponseDto> getSimplePostResponseDtos() {
-        return simplePostResponseDtos;
-    }
-
-    public int getStartPage() {
-        return startPage;
-    }
-
-    public int getEndPage() {
-        return endPage;
-    }
-
-    public boolean isPrev() {
-        return prev;
-    }
-
-    public boolean isNext() {
-        return next;
     }
 }
