+++ backend/src/main/java/com/spring/blog/post/application/dto/response/PostResponseDto.java
@@ -3,7 +3,16 @@ package com.spring.blog.post.application.dto.response;
 import com.spring.blog.post.domain.Post;
 import java.time.LocalDateTime;
 import java.util.List;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class PostResponseDto {
 
     private Long id;
@@ -14,29 +23,6 @@ public class PostResponseDto {
     private Long viewCounts;
     private LocalDateTime createdDate;
     private LocalDateTime modifiedDate;
-
-    private PostResponseDto() {
-    }
-
-    public PostResponseDto(
-        Long id,
-        String title,
-        String content,
-        List<String> urls,
-        String author,
-        Long viewCounts,
-        LocalDateTime createdDate,
-        LocalDateTime modifiedDate
-    ) {
-        this.id = id;
-        this.title = title;
-        this.content = content;
-        this.urls = urls;
-        this.author = author;
-        this.viewCounts = viewCounts;
-        this.createdDate = createdDate;
-        this.modifiedDate = modifiedDate;
-    }
 
     public static PostResponseDto from(Post post) {
         return new PostResponseDto(
@@ -49,37 +35,5 @@ public class PostResponseDto {
             post.getCreatedDate(),
             post.getModifiedDate()
         );
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public String getContent() {
-        return content;
-    }
-
-    public List<String> getUrls() {
-        return urls;
-    }
-
-    public String getAuthor() {
-        return author;
-    }
-
-    public Long getViewCounts() {
-        return viewCounts;
-    }
-
-    public LocalDateTime getCreatedDate() {
-        return createdDate;
-    }
-
-    public LocalDateTime getModifiedDate() {
-        return modifiedDate;
     }
 }
