+++ backend/src/main/java/com/spring/blog/post/infrastructure/dto/FilesResponse.java
@@ -1,19 +1,17 @@
 package com.spring.blog.post.infrastructure.dto;
 
 import java.util.List;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class FilesResponse {
 
     private List<String> urls;
-
-    private FilesResponse() {
-    }
-
-    public FilesResponse(List<String> urls) {
-        this.urls = urls;
-    }
-
-    public List<String> getUrls() {
-        return urls;
-    }
 }
