+++ backend/src/main/java/com/spring/blog/post/presentation/dto/response/PostListResponse.java
@@ -3,7 +3,16 @@ package com.spring.blog.post.presentation.dto.response;
 import com.spring.blog.post.application.dto.response.PostListResponseDto;
 import java.util.List;
 import java.util.stream.Collectors;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class PostListResponse {
 
     private List<SimplePostResponse> simplePostResponses;
@@ -11,23 +20,6 @@ public class PostListResponse {
     private int endPage;
     private boolean prev;
     private boolean next;
-
-    public PostListResponse() {
-    }
-
-    public PostListResponse(
-        List<SimplePostResponse> simplePostResponses,
-        int startPage,
-        int endPage,
-        boolean prev,
-        boolean next
-    ) {
-        this.simplePostResponses = simplePostResponses;
-        this.startPage = startPage;
-        this.endPage = endPage;
-        this.prev = prev;
-        this.next = next;
-    }
 
     public static PostListResponse from(PostListResponseDto postListResponseDto) {
         List<SimplePostResponse> simplePostResponses = postListResponseDto.getSimplePostResponseDtos()
@@ -41,25 +33,5 @@ public class PostListResponse {
             postListResponseDto.isPrev(),
             postListResponseDto.isNext()
         );
-    }
-
-    public List<SimplePostResponse> getSimplePostResponses() {
-        return simplePostResponses;
-    }
-
-    public int getStartPage() {
-        return startPage;
-    }
-
-    public int getEndPage() {
-        return endPage;
-    }
-
-    public boolean getPrev() {
-        return prev;
-    }
-
-    public boolean getNext() {
-        return next;
     }
 }
