+++ backend/src/main/java/com/spring/blog/post/presentation/dto/response/PostResponse.java
@@ -3,7 +3,16 @@ package com.spring.blog.post.presentation.dto.response;
 import com.spring.blog.post.application.dto.response.PostResponseDto;
 import java.time.LocalDateTime;
 import java.util.List;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class PostResponse {
 
     private Long id;
@@ -14,30 +23,6 @@ public class PostResponse {
     private Long viewCounts;
     private LocalDateTime createdDate;
     private LocalDateTime modifiedDate;
-
-
-    private PostResponse() {
-    }
-
-    public PostResponse(
-        Long id,
-        String title,
-        String content,
-        List<String> urls,
-        String author,
-        Long viewCounts,
-        LocalDateTime createdDate,
-        LocalDateTime modifiedDate
-    ) {
-        this.id = id;
-        this.title = title;
-        this.content = content;
-        this.urls = urls;
-        this.author = author;
-        this.viewCounts = viewCounts;
-        this.createdDate = createdDate;
-        this.modifiedDate = modifiedDate;
-    }
 
     public static PostResponse from(PostResponseDto postResponseDto) {
         return new PostResponse(
@@ -50,37 +35,5 @@ public class PostResponse {
             postResponseDto.getCreatedDate(),
             postResponseDto.getModifiedDate()
         );
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public String getContent() {
-        return content;
-    }
-
-    public List<String> getUrls() {
-        return urls;
-    }
-
-    public String getAuthor() {
-        return author;
-    }
-
-    public Long getViewCounts() {
-        return viewCounts;
-    }
-
-    public LocalDateTime getCreatedDate() {
-        return createdDate;
-    }
-
-    public LocalDateTime getModifiedDate() {
-        return modifiedDate;
     }
 }
